requirements.txt
pandas
django_heroku
oauth2client
gunicorn
gspread
datetime
